"Base" Collision Point - Progress to new collision concepts
- Note!: added alot of hard coded integers that need to be figured into variables to make it easier to add sprites and get their collision right (see initCollisionShape)
- Added fields to Renderable to start to think about deliniating between rendering coords and collision coords
- Added the concept of a 'base' collision point which serves to center the collision area of mobs and areaobjects to be closer to the ground, near their visually appearing ground contact
- Refactored collision circles to be ovals, and moved them to be centered at the sprites 'base' collision point
- added some kind of proto mini-map to Area/Mob/AO rendering to help get collision right... visible when Collision Geometry is being shown
- Tried to standardize where Mobs and AOs are positioned and rendered to make collisions and visual interactions more simple to reason about
- Updated target click calculations to work with new figures
- Disabled action window pop-up for now

+++ src/net/wforbes/omnia/overworld/entity/Enemy.java
@@ -12,7 +12,7 @@ public class Enemy extends Mob {
     public void init() {
 
     }
-    public void init(int xPos, int yPos) {}
+    public void init(double xPos, double yPos) {}
 
     @Override
     public void update() {
